Slight restructurization of local-staking contract to fit sylvia better, making storage items public
