+++ contracts/native-staking/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod contract;
 pub mod error;
+mod local_staking_api;
 mod msg;
+mod native_slashig_callback;
 mod state;
 
 #[cfg(not(any(feature = "library", tarpaulin_include)))]
